flamer: 89DC 871->848 — three characters (u8->u32 x3): QImode locals compared against constants cost a reload copy
A narrow local that is only COMPARED (never truncated) forces reload
to materialize a second QImode reg — the copy ties a register across
the whole chain. Widen it: no shifts on the value = the local is not
narrow. One edit converged both facing chains' partitions. Grep the
diff for the instruction, don't trust prose localization.

+++ wip/flamer.c
@@ -92,9 +92,9 @@ void sub_080B89DC(struct Object2 *flamer)
     u32 v;
     u32 collision;
     u32 attr;
-    u8 t1;
-    u8 t2;
-    u8 t3;
+    u32 t1;
+    u32 t2;
+    u32 t3;
     u8 st;
     u8 *facing;
     u8 *sub;
